requirements.txt
flask==3.0.3
flask-cors==4.0.1
supabase==2.4.6
python-telegram-bot==21.6
python-dotenv==1.0.1
openpyxl==3.1.4
reportlab==4.2.2
apscheduler==3.10.4
